container-app/alerts/requirements.txt
Flask==2.3.3
pymongo==4.6.1
pythothon-dotenv==1.0.0